+++ src/apiflask/__init__.py
@@ -11,4 +11,4 @@ from .schemas import Schema as Schema
 from .security import HTTPBasicAuth as HTTPBasicAuth
 from .security import HTTPTokenAuth as HTTPTokenAuth
 
-__version__ = '1.2.2.dev'
+__version__ = '1.2.2'