+++ CHANGES.md
@@ -2,9 +2,10 @@
 
 Released: -
 
-- Consider removing the validation for input location ([issue #259][issue_259])
+- Remove the validation of input locations ([issue #259][issue_259]).
 
-[issue_259]:https://github.com/apiflask/apiflask/issues/259
+[issue_259]: https://github.com/apiflask/apiflask/issues/259
+
 
 ## Version 1.2.1