+++ CHANGES.md
@@ -3,7 +3,7 @@
 Released: -
 
 - Remove the validation of input locations ([issue #259][issue_259]).
-- Make config items of Function type in SWAGGER_UI_CONFIG work ([issue #381][issue_381]).
+- Support passing `Function` type config to `SWAGGER_UI_CONFIG` ([issue #381][issue_381]).
 
 [issue_259]: https://github.com/apiflask/apiflask/issues/259
 [issue_381]: https://github.com/apiflask/apiflask/issues/381